formatting folders and files names
