+++ frontend/src/pages/GamePage/GamePage.jsx
@@ -1,10 +1,10 @@
 import React, { useState, useEffect } from "react";
-import StopWatch from "../components/StopWatch/StopWatch";
-import Countdown from "../components/Countdown/Countdown";
-import "./Game.scss";
-import ShowPictures from "../components/ShowPictures/ShowPictures";
-import Level from "../components/Level/Level";
-import Score from "../components/Score/Score";
+import StopWatch from "../../components/StopWatch/StopWatch";
+import Countdown from "../../components/Countdown/Countdown";
+import "./GamePage.scss";
+import ShowPictures from "../../components/ShowPictures/ShowPictures";
+import Level from "../../components/Level/Level";
+import Score from "../../components/Score/Score";
 
 function Game() {
   const [showComponent, setShowComponent] = useState(false);
