+++ frontend/src/pages/GamePage/GamePage.scss
