+++ frontend/src/pages/RulesPage/RulesPage.scss
